adding some practice on get location from string and show it on map
